+++ src/tt_dlp/__init__.py
@@ -1,3 +1,3 @@
 """tt-dlp package."""
 
-__version__ = "1.3.0"
+__version__ = "1.3.1"